requirements.txt
# Data & ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Database
sqlalchemy==2.0.23

# Visualization & Dashboard
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
streamlit==1.28.2

# Utilities
python-dotenv==1.0.0
geopy==2.4.1
schedule==1.2.1

# Parquet support (for hoopR files in Phase 2)
pyarrow==14.0.1
